backend/requirements.txt
Flask
pymongo
flask-cors
dataclasses
python-dotenv